+++ src/cmd/compile/testdata/script/script_test_basics.txt
@@ -4,6 +4,7 @@
 go build
 [!cgo] skip
 cc -c testdata/mumble.c
+[GOEXPERIMENT:fieldtrack] help exec
 
 -- go.mod --
 module main
